requirements/production.txt
asgiref==3.2.3
boto3==1.11.0
botocore==1.14.0
certifi==2019.11.28
chardet==3.0.4
Django==3.0.3
docutils==0.15.2
idna==2.8
jmespath==0.9.4
PyJWT==1.7.1
python-dateutil==2.8.1
pytz==2019.3
requests==2.22.0
s3transfer==0.3.0
six==1.13.0
sqlparse==0.3.0
urllib3==1.25.7
psycopg2-binary==2.8.4
whitenoise==5.0.1
social-auth-app-django==3.1.0
django-social-share==1.3.2
requests==2.22.0
